requirements.txt
bitarray==2.8.2
blinker==1.6.3
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pymongo==4.5.0
pytest==7.4.2
Werkzeug==3.0.0
requests==2.31.0
python-dotenv==1.0.0
